+++ tsconfig.paths.json
@@ -6,6 +6,7 @@
       "@apis/*": ["src/apis/*"],
       "@images/*": ["src/assets/images/*"],
       "@components/*": ["src/components/*"],
+      "@hooks/*": ["src/hooks/*"],
       "@pages/*": ["src/pages/*"],
       "@routes/*": ["src/routes/*"],
       "@stores/*": ["src/stores/*"],
